+++ src/main/com/marginallyclever/makelangeloRobot/loadAndSave/LoadAndSaveScratch2.java
@@ -42,7 +42,7 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 	
 	private class ScratchVariable {
 		public String name;
-		public float value;
+		public double value;
 
 		public ScratchVariable(String arg0,float arg1) {
 			name=arg0;
@@ -51,11 +51,11 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 	};
 	private class ScratchList {
 		public String name;
-		public ArrayList<Float> contents;
+		public ArrayList<Double> contents;
 
 		public ScratchList(String _name) {
 			name=_name;
-			contents=new ArrayList<Float>();
+			contents=new ArrayList<Double>();
 		}
 	};
 
@@ -269,10 +269,10 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 				ListIterator<?> scriptIter = arr.listIterator();
 				while(scriptIter.hasNext()) {
 					Object varValue = scriptIter.next();
-					float value;
+					double value;
 					if(varValue instanceof Number) {
 						Number num = (Number)varValue;
-						value = (float)num.doubleValue();
+						value = num.doubleValue();
 						list.contents.add(value);
 					} else if(varValue instanceof String) {
 						try {
@@ -370,7 +370,7 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 					// "append:toList:", new value, list name 
 					Object o2 = (Object)scriptIter.next();
 					Object o3 = (Object)scriptIter.next();
-					float value = resolveValue(o2);
+					double value = resolveValue(o2);
 					scratchLists.get(getListID(o3)).contents.add(value);
 				} else if(name.equals("deleteLine:ofList:")) {
 					// "deleteLine:ofList:", index, list name 
@@ -383,7 +383,7 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 					Object o4 = (Object)scriptIter.next();
 					Object o2 = (Object)scriptIter.next();
 					Object o3 = (Object)scriptIter.next();
-					float newValue = resolveValue(o4);
+					double newValue = resolveValue(o4);
 					int listIndex = (int)resolveListIndex(o2,o3);
 					scratchLists.get(getListID(o3)).contents.add(listIndex,newValue);
 				} else if(name.equals("setLine:ofList:to:")) {
@@ -391,13 +391,13 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 					Object o4 = (Object)scriptIter.next();
 					Object o2 = (Object)scriptIter.next();
 					Object o3 = (Object)scriptIter.next();
-					float newValue = resolveValue(o4);
+					double newValue = resolveValue(o4);
 					int listIndex = (int)resolveListIndex(o2,o3);
 					scratchLists.get(getListID(o3)).contents.set(listIndex,newValue);
 				} else if(name.equals("wait:elapsed:from:")) {
 					// dwell - does nothing.
 					Object o2 = (Object)scriptIter.next();
-					float seconds = resolveValue(o2);
+					double seconds = resolveValue(o2);
 					System.out.println("dwell "+seconds+" seconds.");
 					continue;
 				} else if(name.equals("putPenUp")) {
@@ -411,9 +411,9 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 					System.out.println("pen down");
 				} else if(name.equals("gotoX:y:")) {
 					Object o2 = (Object)scriptIter.next();
-					float x = resolveValue(o2);
+					double x = resolveValue(o2);
 					Object o3 = (Object)scriptIter.next();
-					float y = resolveValue(o3);
+					double y = resolveValue(o3);
 					
 					turtle.setX(x);
 					turtle.setY(y);
@@ -421,47 +421,47 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 					System.out.println("Move to ("+turtle.getX()+","+turtle.getY()+")");
 				} else if(name.equals("changeXposBy:")) {
 					Object o2 = (Object)scriptIter.next();
-					float v = resolveValue(o2);
+					double v = resolveValue(o2);
 					turtle.setX(turtle.getX()+v);
 					this.moveTo(out, turtle.getX(), turtle.getY(), penUp);
 					//System.out.println("Move to ("+turtle.getX()+","+turtle.getY()+")");
 				} else if(name.equals("changeYposBy:")) {
 					Object o2 = (Object)scriptIter.next();
-					float v = resolveValue(o2);
+					double v = resolveValue(o2);
 					turtle.setY(turtle.getY()+v);
 					this.moveTo(out, turtle.getX(), turtle.getY(), penUp);
 					//System.out.println("Move to ("+turtle.getX()+","+turtle.getY()+")");
 				} else if(name.equals("forward:")) {
 					Object o2 = (Object)scriptIter.next();
-					float v = resolveValue(o2);
+					double v = resolveValue(o2);
 					turtle.move(v);
 					this.moveTo(out, turtle.getX(), turtle.getY(), penUp);
 					System.out.println("Move forward "+v+" mm");
 				} else if(name.equals("turnRight:")) {
 					Object o2 = (Object)scriptIter.next();
-					float degrees = resolveValue(o2);
+					double degrees = resolveValue(o2);
 					turtle.turn(-degrees);
 					System.out.println("Right "+degrees+" degrees.");
 				} else if(name.equals("turnLeft:")) {
 					Object o2 = (Object)scriptIter.next();
-					float degrees = resolveValue(o2);
+					double degrees = resolveValue(o2);
 					turtle.turn(degrees);
 					System.out.println("Left "+degrees+" degrees.");
 				} else if(name.equals("xpos:")) {
 					Object o2 = (Object)scriptIter.next();
-					float v = resolveValue(o2);
+					double v = resolveValue(o2);
 					turtle.setX(v);
 					this.moveTo(out, turtle.getX(), turtle.getY(), penUp);
 					//System.out.println("Move to ("+turtle.getX()+","+turtle.getY()+")");
 				} else if(name.equals("ypos:")) {
 					Object o2 = (Object)scriptIter.next();
-					float v = resolveValue(o2);
+					double v = resolveValue(o2);
 					turtle.setY(v);
 					this.moveTo(out, turtle.getX(), turtle.getY(), penUp);
 					//System.out.println("Move to ("+turtle.getX()+","+turtle.getY()+")");
 				} else if(name.equals("heading:")) {
 					Object o2 = (Object)scriptIter.next();
-					float degrees = resolveValue(o2);
+					double degrees = resolveValue(o2);
 					turtle.setAngle(degrees);
 					//System.out.println("Turn to "+degrees);
 				} else if(name.equals("setVar:to:")) {
@@ -533,22 +533,22 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 		if(name.equals(">")) {
 			Object o2 = (Object)scriptIter.next();
 			Object o3 = (Object)scriptIter.next();
-			float a = resolveValue(o2);
-			float b = resolveValue(o3);
+			double a = resolveValue(o2);
+			double b = resolveValue(o3);
 			return a > b;
 		}
 		if(name.equals("<")) {
 			Object o2 = (Object)scriptIter.next();
 			Object o3 = (Object)scriptIter.next();
-			float a = resolveValue(o2);
-			float b = resolveValue(o3);
+			double a = resolveValue(o2);
+			double b = resolveValue(o3);
 			return a < b;
 		}
 		if(name.equals("=")) {
 			Object o2 = (Object)scriptIter.next();
 			Object o3 = (Object)scriptIter.next();
-			float a = resolveValue(o2);
-			float b = resolveValue(o3);
+			double a = resolveValue(o2);
+			double b = resolveValue(o3);
 			return a == b; 
 		}
 		if(name.equals("not")) {
@@ -575,7 +575,7 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 	 * @return the calculated final value.
 	 * @throws Exception
 	 */
-	private float resolveValue(Object obj) throws Exception {
+	private double resolveValue(Object obj) throws Exception {
 		if(obj instanceof String) {
 			// probably a variable
 			String firstName = obj.toString();
@@ -600,7 +600,7 @@ public class LoadAndSaveScratch2 extends ImageManipulator implements LoadAndSave
 		
 		if(obj instanceof Number) {
 			Number num = (Number)obj;
-			return (float)num.doubleValue();
+			return (double)num.doubleValue();
 		}
 		
 		if(obj instanceof JSONArray) {
